Fix solution: java/bird-watcher
Convert the `long` datetype to `int`

+++ java/bird-watcher/src/main/java/BirdWatcher.java
@@ -28,6 +28,6 @@ class BirdWatcher {
     }
 
     public int getBusyDays() {
-        return Arrays.stream(birdsPerDay).filter(n -> n >= 5).count();
+        return (int) Arrays.stream(birdsPerDay).filter(n -> n >= 5).count();
     }
 }
